한자 or 이모지 단축키 -> 한자 and 이모지 단축키
