+++ Preferences/GureumPreferencePane.swift
@@ -1,0 +1,18 @@
+//
+//  GureumPreferencePane.swift
+//  Preferences
+//
+//  Created by Presto on 04/10/2019.
+//  Copyright © 2019 youknowone.org. All rights reserved.
+//
+
+import Cocoa
+import PreferencePanes
+
+final class GureumPreferencePane: NSPreferencePane {
+    @IBOutlet private var viewController: NSViewController!
+
+    override func mainViewDidLoad() {
+        super.mainViewDidLoad()
+    }
+}
